Handle a null VCS Root
Following the removal of the StartProcessor, it was found that the
BuildInformationProcess would fail in the event that no vcsRoot
was associated with the build configuration.

A much reduced StartProcessor as been re-added to determine vcsType
(unable to be performed from agent-side) - and the
BuildInformationBuildProcess will serialise a null vcsRoot
into the bulidInformation file passed to octo.exe.